added tests for vocab and tokenizer and renamed file
